Some more work on reflection
